Update Web Cloaker Editor
To Insert code into the browser. Make a bookmark, click more copy the code into the link bar, Change the name (to something cool)

When you want to cloak a tab just click on the bookmark!

Simple! when wanting to change code go to any javascript coding website and follow in instructions prebuilt into the code.